reorganize integration specs files
